jadl-trafo: long -> int
git-svn-id: https://cvs.dai-labor.de/jiac/toolipse3/trunk@31307 d8bac29e-b7f5-0310-9bdb-a06bcefdf3b8

+++ de.dailab.vsdt.trafo.jiacv/src/de/dailab/vsdt/trafo/jiacv/export/Bpmn2JiacVValidation.java
@@ -27,7 +27,7 @@ public class Bpmn2JiacVValidation extends DefaultBpmnValidation {
 		basicDataTypeMap.put(Util.TYPE_STRING, "string");
 		basicDataTypeMap.put(Util.TYPE_FLOAT, "float");
 		basicDataTypeMap.put(Util.TYPE_DOUBLE, "double");
-		basicDataTypeMap.put(Util.TYPE_LONG, "long");
+		basicDataTypeMap.put(Util.TYPE_LONG, "int");
 	}
 	
 	/**
